move grafana templating to another pkg
