+++ pkg/grafana/grafana.go
@@ -1,38 +1,20 @@
 package grafana
 
 import (
-	"bytes"
 	"context"
 	_ "embed"
 	"fmt"
-	"html/template"
 	"strings"
 
 	"github.com/grafana/grafana-openapi-client-go/client"
 	"github.com/grafana/grafana-openapi-client-go/models"
 	"github.com/pkg/errors"
 	"sigs.k8s.io/controller-runtime/pkg/log"
-
-	"github.com/giantswarm/observability-operator/api/v1alpha1"
 )
 
 const (
 	SharedOrgName = "Shared Org"
-
-	grafanaAdminRole  = "Admin"
-	grafanaEditorRole = "Editor"
-	grafanaViewerRole = "Viewer"
 )
-
-var (
-	//go:embed templates/grafana-user-values.yaml.template
-	grafanaUserConfig         string
-	grafanaUserConfigTemplate *template.Template
-)
-
-func init() {
-	grafanaUserConfigTemplate = template.Must(template.New("grafana-user-values.yaml").Parse(grafanaUserConfig))
-}
 
 func CreateOrganization(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, organization Organization) (Organization, error) {
 	logger := log.FromContext(ctx)
@@ -62,7 +44,7 @@ func UpdateOrganization(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, 
 	logger := log.FromContext(ctx)
 
 	logger.Info("updating organization")
-	found, err := findByID(ctx, grafanaAPI, organization.ID)
+	found, err := findByID(grafanaAPI, organization.ID)
 	if err != nil {
 		if isNotFound(err) {
 			logger.Info("organization id not found, creating")
@@ -105,7 +87,7 @@ func DeleteByID(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, id int64
 	logger := log.FromContext(ctx)
 
 	logger.Info("deleting organization")
-	_, err := findByID(ctx, grafanaAPI, id)
+	_, err := findByID(grafanaAPI, id)
 	if err != nil {
 		logger.Error(err, fmt.Sprintf("failed to find organization with ID: %d", id))
 	}
@@ -118,47 +100,6 @@ func DeleteByID(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, id int64
 	logger.Info("deleted organization")
 
 	return nil
-}
-
-func GenerateGrafanaConfiguration(organizations []v1alpha1.GrafanaOrganization) (string, error) {
-	var orgMappings []string
-	orgMappings = append(orgMappings, fmt.Sprintf(`"*:%s:%s"`, SharedOrgName, grafanaAdminRole))
-	for _, organization := range organizations {
-		rbac := organization.Spec.RBAC
-		organizationName := organization.Spec.DisplayName
-		for _, adminOrgAttribute := range rbac.Admins {
-			orgMappings = append(orgMappings, buildOrgMapping(organizationName, adminOrgAttribute, grafanaAdminRole))
-		}
-		for _, editorOrgAttribute := range rbac.Editors {
-			orgMappings = append(orgMappings, buildOrgMapping(organizationName, editorOrgAttribute, grafanaEditorRole))
-		}
-		for _, viewerOrgAttribute := range rbac.Viewers {
-			orgMappings = append(orgMappings, buildOrgMapping(organizationName, viewerOrgAttribute, grafanaViewerRole))
-		}
-	}
-
-	orgMapping := strings.Join(orgMappings, " ")
-
-	data := struct {
-		OrgMapping string
-	}{
-		OrgMapping: orgMapping,
-	}
-
-	var values bytes.Buffer
-	err := grafanaUserConfigTemplate.Execute(&values, data)
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-
-	return values.String(), nil
-}
-
-func buildOrgMapping(organizationName, userOrgAttribute, role string) string {
-	// We need to escape the colon in the userOrgAttribute
-	u := strings.ReplaceAll(userOrgAttribute, ":", "\\:")
-	// We add double quotes to the org mapping to support spaces in display names
-	return fmt.Sprintf(`"%s:%s:%s"`, u, organizationName, role)
 }
 
 func isNotFound(err error) bool {
@@ -174,7 +115,7 @@ func isNotFound(err error) bool {
 func assertNameIsAvailable(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, organization Organization) error {
 	logger := log.FromContext(ctx)
 
-	found, err := findByName(ctx, grafanaAPI, organization.Name)
+	found, err := findByName(grafanaAPI, organization.Name)
 	// We only error if we have any error other than a 404
 	if err != nil && !isNotFound(err) {
 		logger.Error(err, fmt.Sprintf("failed to find organization with name: %s", organization.Name))
@@ -189,7 +130,7 @@ func assertNameIsAvailable(ctx context.Context, grafanaAPI *client.GrafanaHTTPAP
 }
 
 // findByName is a wrapper function used to find a Grafana organization by its name
-func findByName(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, name string) (*Organization, error) {
+func findByName(grafanaAPI *client.GrafanaHTTPAPI, name string) (*Organization, error) {
 	organization, err := grafanaAPI.Orgs.GetOrgByName(name)
 	if err != nil {
 		return nil, errors.WithStack(err)
@@ -202,7 +143,7 @@ func findByName(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, name str
 }
 
 // findByID is a wrapper function used to find a Grafana organization by its id
-func findByID(ctx context.Context, grafanaAPI *client.GrafanaHTTPAPI, orgID int64) (*Organization, error) {
+func findByID(grafanaAPI *client.GrafanaHTTPAPI, orgID int64) (*Organization, error) {
 	organization, err := grafanaAPI.Orgs.GetOrgByID(orgID)
 	if err != nil {
 		return nil, errors.WithStack(err)
